backend/requirements.txt
openai-whisper
soundfile
numpy
flask
